refactor(setup-wallet): renaming
